refactor(frontend): inline and delete the empty AuthShell wrapper
